Build fluentbit from git repo, latest master commit at the time
Summary:
Adds ability to build from any commit in git repo, in addition to building from a specific release.

Upstart does not expose env variables to jobs launched by it, therefore we need to explicitly set `HOSTNAME`. Since it's value is not static, it has to be wrapped in `script` block.

Related: T7.

Test Plan:
Installed resulting *.deb package on tester and verified that it works.

```
[2017/07/10 09:33:39] [ info] [engine] started
[0] cpu.local: [1499679229.001839108, {"cpu_p"=>0.000000, "user_p"=>0.000000, "system_p"=>0.000000, "cpu0.p_cpu"=>0.000000, "cpu0.p_user"=>0.000000, "cpu0.p_system"=>0.000000, "host"=>"vagrant-ubuntu-trusty-64", "aa"=>"cc"}]
```

Reviewers: ivan.vergiliev

Reviewed By: ivan.vergiliev

Differential Revision: https://heapanalytics.phacility.com/D12